+++ src/pocketmine/level/biome/Biome.php
@@ -21,20 +21,10 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\level\generator\biome;
+namespace pocketmine\level\biome;
 
 use pocketmine\block\Block;
 use pocketmine\level\ChunkManager;
-use pocketmine\level\generator\normal\biome\DesertBiome;
-use pocketmine\level\generator\normal\biome\ForestBiome;
-use pocketmine\level\generator\normal\biome\IcePlainsBiome;
-use pocketmine\level\generator\normal\biome\MountainsBiome;
-use pocketmine\level\generator\normal\biome\OceanBiome;
-use pocketmine\level\generator\normal\biome\PlainBiome;
-use pocketmine\level\generator\normal\biome\RiverBiome;
-use pocketmine\level\generator\normal\biome\SmallMountainsBiome;
-use pocketmine\level\generator\normal\biome\SwampBiome;
-use pocketmine\level\generator\normal\biome\TaigaBiome;
 use pocketmine\level\generator\populator\Populator;
 use pocketmine\utils\Random;
