+++ src/pocketmine/level/biome/DesertBiome.php
@@ -21,33 +21,20 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\level\generator\normal\biome;
+namespace pocketmine\level\biome;
 
-use pocketmine\block\Sapling;
-use pocketmine\level\generator\populator\TallGrass;
-use pocketmine\level\generator\populator\Tree;
 
-class TaigaBiome extends SnowyBiome{
+class DesertBiome extends SandyBiome{
 
 	public function __construct(){
 		parent::__construct();
+		$this->setElevation(63, 74);
 
-		$trees = new Tree(Sapling::SPRUCE);
-		$trees->setBaseAmount(10);
-		$this->addPopulator($trees);
-
-		$tallGrass = new TallGrass();
-		$tallGrass->setBaseAmount(1);
-
-		$this->addPopulator($tallGrass);
-
-		$this->setElevation(63, 81);
-
-		$this->temperature = 0.05;
-		$this->rainfall = 0.8;
+		$this->temperature = 2;
+		$this->rainfall = 0;
 	}
 
 	public function getName() : string{
-		return "Taiga";
+		return "Desert";
 	}
 }
