+++ src/pocketmine/level/biome/GrassyBiome.php
@@ -21,27 +21,20 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\level\generator\normal\biome;
+namespace pocketmine\level\biome;
 
-use pocketmine\level\generator\populator\TallGrass;
+use pocketmine\block\Block;
+use pocketmine\block\BlockFactory;
 
-class OceanBiome extends GrassyBiome{
+abstract class GrassyBiome extends NormalBiome{
 
 	public function __construct(){
-		parent::__construct();
-
-		$tallGrass = new TallGrass();
-		$tallGrass->setBaseAmount(5);
-
-		$this->addPopulator($tallGrass);
-
-		$this->setElevation(46, 58);
-
-		$this->temperature = 0.5;
-		$this->rainfall = 0.5;
-	}
-
-	public function getName() : string{
-		return "Ocean";
+		$this->setGroundCover([
+			BlockFactory::get(Block::GRASS, 0),
+			BlockFactory::get(Block::DIRT, 0),
+			BlockFactory::get(Block::DIRT, 0),
+			BlockFactory::get(Block::DIRT, 0),
+			BlockFactory::get(Block::DIRT, 0),
+		]);
 	}
 }
