+++ src/pocketmine/level/biome/HellBiome.php
@@ -21,20 +21,11 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\level\generator\normal\biome;
+namespace pocketmine\level\biome;
 
-class SwampBiome extends GrassyBiome{
-
-	public function __construct(){
-		parent::__construct();
-
-		$this->setElevation(62, 63);
-
-		$this->temperature = 0.8;
-		$this->rainfall = 0.9;
-	}
+class HellBiome extends Biome{
 
 	public function getName() : string{
-		return "Swamp";
+		return "Hell";
 	}
 }
