+++ src/pocketmine/level/biome/NormalBiome.php
@@ -21,27 +21,8 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\level\generator\normal\biome;
+namespace pocketmine\level\biome;
 
-use pocketmine\level\generator\populator\TallGrass;
+abstract class NormalBiome extends Biome{
 
-class IcePlainsBiome extends SnowyBiome{
-
-	public function __construct(){
-		parent::__construct();
-
-		$tallGrass = new TallGrass();
-		$tallGrass->setBaseAmount(5);
-
-		$this->addPopulator($tallGrass);
-
-		$this->setElevation(63, 74);
-
-		$this->temperature = 0.05;
-		$this->rainfall = 0.8;
-	}
-
-	public function getName() : string{
-		return "Ice Plains";
-	}
 }
