+++ src/pocketmine/level/biome/PlainBiome.php
@@ -21,27 +21,27 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\level\generator\normal\biome;
+namespace pocketmine\level\biome;
 
 use pocketmine\level\generator\populator\TallGrass;
 
-class RiverBiome extends GrassyBiome{
+class PlainBiome extends GrassyBiome{
 
 	public function __construct(){
 		parent::__construct();
 
 		$tallGrass = new TallGrass();
-		$tallGrass->setBaseAmount(5);
+		$tallGrass->setBaseAmount(12);
 
 		$this->addPopulator($tallGrass);
 
-		$this->setElevation(58, 62);
+		$this->setElevation(63, 74);
 
-		$this->temperature = 0.5;
-		$this->rainfall = 0.7;
+		$this->temperature = 0.8;
+		$this->rainfall = 0.4;
 	}
 
 	public function getName() : string{
-		return "River";
+		return "Plains";
 	}
 }
