+++ src/pocketmine/level/biome/SmallMountainsBiome.php
@@ -21,10 +21,18 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\level\generator\normal\biome;
+namespace pocketmine\level\biome;
 
-use pocketmine\level\generator\biome\Biome;
 
-abstract class NormalBiome extends Biome{
+class SmallMountainsBiome extends MountainsBiome{
 
+	public function __construct(){
+		parent::__construct();
+
+		$this->setElevation(63, 97);
+	}
+
+	public function getName() : string{
+		return "Small Mountains";
+	}
 }
