+++ src/pocketmine/level/biome/SwampBiome.php
@@ -21,27 +21,20 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\level\generator\normal\biome;
+namespace pocketmine\level\biome;
 
-use pocketmine\level\generator\populator\TallGrass;
-
-class PlainBiome extends GrassyBiome{
+class SwampBiome extends GrassyBiome{
 
 	public function __construct(){
 		parent::__construct();
 
-		$tallGrass = new TallGrass();
-		$tallGrass->setBaseAmount(12);
-
-		$this->addPopulator($tallGrass);
-
-		$this->setElevation(63, 74);
+		$this->setElevation(62, 63);
 
 		$this->temperature = 0.8;
-		$this->rainfall = 0.4;
+		$this->rainfall = 0.9;
 	}
 
 	public function getName() : string{
-		return "Plains";
+		return "Swamp";
 	}
 }
